+++ CHANGELOG.md
@@ -10,6 +10,8 @@ Versioning].
 ## [Unreleased]
 
 - Fix bug where rules following a wildcard would not be matched.
+- Fix matching deprecations for `npm:` aliased dependencies.
+- Improve matching of aliased dependencies in deprecation paths.
 - Let `*` match 0-or-more instead of 1-or-more.
 
 ## [0.3.0] - 2024-11-03